server/requirements.txt
Django~=4.2
djangorestframework==3.14.0
django-filter~=23.1
django-environ~=0.10.0
psycopg2-binary~=2.9.6
gunicorn~=20.1.0
drf-yasg~=1.21.4
pika~=1.3.1
